Move logic of create colors from App file to useColorScheme.js
