+++ src/api/groqClient.js
@@ -1,0 +1,8 @@
+import Groq, { NotFoundError } from "groq-sdk";
+
+const groq = new Groq({
+  apiKey: import.meta.env.VITE_GROQ_API_KEY,
+  dangerouslyAllowBrowser: true, //Better to move in server part
+});
+
+export default groq;
